Create +1 831 275 1449 JOIN THE ILLUMINATI BROTHERHOOD IMMEDIATELY IN SOUTH SUDAN,JOHANNESBURG ,SOWETO
Who were the Illuminati, and do they really control the world? Here's what we know about one of history's most alluring secret societies, including how you became a member.

What is the Illuminati?
The Illuminati is a name given to both a real and fictitious society. The latter has fuelled conspiracy theories for years, with people claiming it to be a secretive and mysterious worldwide organisation intent on world domination – as well as being behind some of history’s greatest revolutions and assassinations.

What was the original Illuminati?
The Illuminati was a secret society formed in Bavaria (now part of modern-day Germany) that existed from 1776 to 1785 – its members originally referred to themselves as Perfectibilists.